refactor: move interfaces to the corresponding directories
