+++ src/core/repositories/album/interface/album.interface.ts
@@ -1,0 +1,6 @@
+export interface Album {
+  id: string; // uuid v4
+  name: string;
+  year: number;
+  artistId: string | null; // refers to Artist
+}
