+++ src/core/repositories/favorites/interface/favorites.interface.ts
@@ -1,0 +1,5 @@
+export interface Favorites {
+  artists: string[]; // favorite artists ids
+  albums: string[]; // favorite albums ids
+  tracks: string[]; // favorite tracks ids
+}
